+++ src/apps/toolbar/modules/main/application.ts
@@ -5,7 +5,7 @@ import { cloneDeep, debounce } from 'lodash';
 
 import { store } from '../shared/store/infra';
 
-export const SavePlaceholderAsCustom = debounce(async () => {
+export const SaveToolbarItems = debounce(async () => {
   const { placeholder } = store.getState();
   const toBeSaved = cloneDeep(placeholder);
   const filePath = await path.join(await path.appDataDir(), 'toolbar_items.yml');