+++ src/apps/toolbar/modules/main/infra.tsx
@@ -16,7 +16,7 @@ import { PowerModule } from '../Power/infra';
 import { SettingsModule } from '../Settings/infra';
 
 import { RootActions, Selectors } from '../shared/store/app';
-import { SavePlaceholderAsCustom } from './application';
+import { SaveToolbarItems } from './application';
 
 import { AnimatedDropdown } from '../../../shared/components/AnimatedWrappers';
 import { useWindowFocusChange } from '../../../shared/hooks';
@@ -129,7 +129,7 @@ export function ToolBar({ structure }: Props) {
     payload = apps.slice(dividerEnd + 1) as ToolbarItem[];
     dispatch(RootActions.setItemsOnRight(payload));
 
-    SavePlaceholderAsCustom()?.catch(console.error);
+    SaveToolbarItems()?.catch(console.error);
   }, []);
 
   const shouldBeHidden =