+++ changelog.md
@@ -13,6 +13,7 @@
 - not showing save button after icon packs change.
 - app failing when powershell is not part of the $PATH enviroment.
 - dock items no updatings paths on store updates.
+- unremovable workspace module on toolbar.
 
 ## [2.0.14]
 ### hotfix